Building a Restful API: Routing Requests
we start off by defining our request router, once we have our routes
we define the handlers that we need to handle directing requests to those routes.

we create a chosenHandler function that runs everytime the 'end' event is emitted.
a handler is chosen if the trimmed path exists in the router, if it does then we
return the router[trimmedPath], (ps:the router is a key:value pair that links
each route to a specific handler). other wise we return the notFound handler.

we then construct our data object to include all the request data that
we want to send to our chosenHandler.
    var data = {
      'trimmedPath': trimmedPath,
      'queryStringObject': queryStringObject,
      'method': method,
      'headers': headers,
      'payload': buffer
    };
once we've gotten our data we parse it to the chosenHandlerFunction,
along with a callback function that accepts a statuscode and payload
as arguments. this is then used to return a response to the request.